Updated fixed.dat for romcal-V6
This file contains the fixed memorials and memorials for the General Calendar as used by Kenneth Bath's romcal program. I have brought it up to date as of 2025 (I had to guess for Bl. Carlo Acutis). Some of these are specific to the United States, so use with care and edit as necessary for your particular location. To use this, replace the included fixed.dat with this one and recompile.